+++ pyproject.toml
@@ -240,6 +240,7 @@ omit = [
     "src/export/license_export_utils.py",
     "src/export/org_admin_exporter.py",
     "src/export/org_alarm_event_exporter.py",
+    "src/export/org_client_security_exporter.py",
     "src/export/org_config_exporter.py",
     "src/export/org_template_exporter.py",
     "src/export/site_config_exporter.py",
